// src/types/Phones.tsx
export interface Phones {
  id: string;
  category: string;
  namespaceId: string;
  name: string;
  capacityAvailable: [];
  capacity: number;
  priceRegular: number;
  priceDiscount: number;
  colorsAvailable: [];
  color: string;
  images: string;
  ram: number;
  screen: number;
  year: number;
}